resolve some exit multithreading problems

+++ pg_dump/pgdump_thread.py
@@ -47,9 +47,9 @@ class PgDumpThread(Thread):
                     "Pgdump thread job started at %s has exceeded max number of retries"
                 )
                 continue
+            path = core.backup_folder_path(self.job.filename)
             try:
                 core.run_pg_dump(self.job.filename)
-                path = core.backup_folder_path(self.job.filename)
                 if path.exists() and not path.stat().st_size:
                     log.error("Error pgdump thread %s: backup file empty", self.number)
                     raise core.CoreSubprocessError()
@@ -58,10 +58,14 @@ class PgDumpThread(Thread):
                 log.error(
                     "Error pgdump thread %s: error performing pgdump", self.number
                 )
-                core.backup_folder_path(self.job.filename).unlink()
-                log.error("Removed empty backup file %s", self.job.filename)
+                if path.exists():
+                    core.backup_folder_path(self.job.filename).unlink()
+                    log.error("Removed empty backup file %s", self.job.filename)
                 self.cooling_period()
                 self.job.retries += 1
+                log.error(
+                    "Adding job back to PGDUMP_QUEUE after error: %s", self.job.filename
+                )
                 core.PGDUMP_QUEUE.put(self.job)
         log.info("Pgdump thread %s has stopped", self.number)
 
@@ -75,13 +79,14 @@ class PgDumpThread(Thread):
             self.number,
             release_time,
         )
-        while self.running:
+        while self.running():
             now = datetime.utcnow()
             if now > release_time:
-                log.info("Pgdump thread %s finished cooling period", self.number)
                 self.cooling = False
+                log.info("Pgdump thread %s finished cooling period", self.number)
                 return
             time.sleep(0.02)
+        log.info("Pgdump thread %s skipping cooling period", self.number)
 
     def stop(self):
         log.info("Stopping pgdump thread %s", self.number)
